+++ noodles-sam/examples/sam_write_async.rs
@@ -27,7 +27,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     for _ in 0..3 {
         let record = sam::Record::default();
-        writer.write_record(&record).await?;
+        writer.write_record(&header, &record).await?;
     }
 
     Ok(())